+++ src/main/kotlin/com/kylecorry/sol/units/EnergyUnits.kt
@@ -1,6 +1,6 @@
 package com.kylecorry.sol.units
 
-enum class EnergyUnits(val id: Int, val joules: Float, val isMetric: Boolean) {
-    Joules(1, 1f, true),
-    FootPounds(2, 1.355818f, false),
+enum class EnergyUnits(val id: Int, val joules: Double, val isMetric: Boolean) {
+    Joules(1, 1.0, true),
+    FootPounds(2, 1.355818, false),
 }
