+++ src/main/kotlin/com/kylecorry/sol/units/PressureUnits.kt
@@ -1,9 +1,9 @@
 package com.kylecorry.sol.units
 
-enum class PressureUnits(val id: Int, val hpa: Float) {
-    Hpa(1, 1f),
-    Mbar(2, 1f),
-    Inhg(3, 1 / 0.02953f),
-    Psi(4, 1 / 0.014503774f),
-    MmHg(5, 1.3332239f)
+enum class PressureUnits(val id: Int, val hpa: Double) {
+    Hpa(1, 1.0),
+    Mbar(2, 1.0),
+    Inhg(3, 1 / 0.02953),
+    Psi(4, 1 / 0.014503774),
+    MmHg(5, 1.3332239)
 }
