+++ src/main/kotlin/com/kylecorry/sol/units/WeightUnits.kt
@@ -1,10 +1,10 @@
 package com.kylecorry.sol.units
 
-enum class WeightUnits(val id: Int, val grams: Float) {
-    Pounds(1, 453.592f),
-    Ounces(2, 28.3495f),
-    Kilograms(3, 1000f),
-    Grams(4, 1f),
-    Milligrams(5, 0.001f),
-    Grains(6, 0.06479891f),
+enum class WeightUnits(val id: Int, val grams: Double) {
+    Pounds(1, 453.592),
+    Ounces(2, 28.3495),
+    Kilograms(3, 1000.0),
+    Grams(4, 1.0),
+    Milligrams(5, 0.001),
+    Grains(6, 0.06479891),
 }
